Add starter files
At this point some functionalities work but there's still more to do.
Now the focus is just adding the project on the NPM registry and test
if the creating of folders/boilerplate code work as expected.

I already did explain some of the functionalities we have so far on my
other repo, you can check it out here:
https://github.com/lindelwa122/odin-restaurant-page