Modify _createElement
I removed the need for attributes as everything can be included inside
options. This is much better because it removes one extra step when it
comes to creating DOM which makes the process, a little easier, and
less error-prone. This also removes a lot of isolated cases inside the
functions where some cases were handled differently.